requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
python-dotenv
groq>=0.5.0
openai>=1.10.0
google-generativeai>=0.4.0
tavily-python
openai-whisper
opencv-python
Pillow
setuptools-rust
chromadb>=0.4.0
sentence-transformers
tiktoken
streamlit
numpy